Multiple Changes
selectedUnits is now cleared when the player moves a unit.

The pathfinding algorithm now ensures that when trying to move diagonally, the tiles above and below are checked to ensure corners are not cut.

UnitGO now updates the occupyingTile and occupyingUnit variables correctly when the destination is taken up by another unit mid pathing.

:chart: