Bug Fixes
Objects that can not be pathed to have their jobs removed so the units do not idle.
Instead of using new Pathfinding() multiple times, A single Pathfinding object has been added to GameManager.
ExecuteJobs() has had its foreach changed to a for loop so that it does not generate an error when a job is removed from the jobs list.
Objects now store a reference to the job that is harvesting it.
Units now have a FreeUnitFromJob function.
Units will no longer stop pathing if another unit steps on its destination tile.